Added controls to change iteration count, radius, and position
After making a value change in the secondary window, the change will not take place until you choose another fractal.

It seems that whatever value you input into the fractal, it will carry over to the next load instead of loading in a fresh fractal.

H works, but K does not when changing the position of the graph for some reason. when H is changed, the radius around the fractal distorts into a strange shape.